requirements.txt
websocket-client==1.7.0
pyModbusTCP==0.2.0
pynmea2==1.19.0
pytest-mock==3.12.0
pytest==7.4.3
satella==2.23.3
setuptools==68.2.2
wheel==0.41.2
pymodbus==3.5.2
paho-mqtt==1.6.1
websockets==12.0